Multiple Nagios concurrency fix This commit fixes a concurrency problem that occurs when calling from multiple nagios servers. It adds the script's PID number to all temporary file names used so that running the script again before any previous runs complete cannot cause a problem.
The problem was most noticeable when a second call to the script
resulted in a collision on the PID lock file check resulting in
the end user seeing the "need to reduce the check duration or
increase the interval between checks" error message even though
the issue was not the duration or frequency of the checks but the
fact that a check from a different nagios server had a scheduling
collision.

It is possible, but highly unlikely, that multiple runs could
become concurrent in a way that would allow the PID lock file
check to pass yet still allow a file use collision that would
produce bad data.